% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{contains}
\alias{contains}
\title{Contains}
\usage{
contains(network, i, j)
}
\arguments{
\item{network}{An instance of the incgraph.network class}

\item{i}{A node in \code{network}}

\item{j}{A node in \code{network}}
}
\value{
\code{TRUE} if the network contains (i, j)
}
\description{
\code{contains} returns \code{TRUE} if the network contains the edge (i, j).
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
