% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{flip}
\alias{flip}
\title{Modify edge}
\usage{
flip(network, i, j)
}
\arguments{
\item{network}{An instance of the incgraph.network class}

\item{i}{A node in \code{network}}

\item{j}{A node in \code{network}}
}
\description{
\code{flip} modifies an edge in the network. If it is contained in the network, it is removed from the network, otherwise it is added to the network.
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
