% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{get.neighbours}
\alias{get.neighbours}
\title{Neighbours}
\usage{
get.neighbours(network, i)
}
\arguments{
\item{network}{An instance of the incgraph.network class}

\item{i}{A node in \code{network}}
}
\value{
Returns all neighbours of node \code{i}
}
\description{
\code{get.neighbours} returns a vector of all neighbours of \code{i}.
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
