% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{network.as.matrix}
\alias{network.as.matrix}
\title{Network as matrix}
\usage{
network.as.matrix(network)
}
\arguments{
\item{network}{An instance of the incgraph.network class}
}
\description{
\code{network.as.matrix} returns the network as a matrix
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
