% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{orca.halfdelta}
\alias{orca.halfdelta}
\title{Modify edge}
\usage{
orca.halfdelta(network, i, j)
}
\arguments{
\item{network}{An instance of the incgraph.network class}

\item{i}{A node in \code{network}}

\item{j}{A node in \code{network}}
}
\description{
\code{orca.halfdelta} calculates the orca counts for a network that has just been changed.
}
