% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{reset}
\alias{reset}
\title{Reset network}
\usage{
reset(network)
}
\arguments{
\item{network}{An instance of the incgraph.network class}
}
\description{
\code{reset} resets all the data structures so that all edges are removed from the network.
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
