% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incgraph_wrapper.R
\name{set.network}
\alias{set.network}
\title{Set a given network to contain the given links}
\usage{
set.network(network, links)
}
\arguments{
\item{network}{An instance of the incgraph.network class}

\item{links}{A matrix with 2 columns and N rows, 1 row for each edge to be loaded in the network}
}
\description{
\code{set.network} sets a given network to contain the given links.
}
\details{
This first resets the network and adds all given links. For minor changes to the network,
the usage of \code{\link{flip}} is recommended.
}
\seealso{
See \code{\link{new.incgraph.network}} for examples and usage.
}
